implemented independent kmax in P(k) and C_l; corrected bugs in error messages leading to warnings during compilation

+++ source/output.c
@@ -55,7 +55,7 @@ int output_init(
 	}
 	
 	fprintf(out[index_ic],"# number of values of l:\n");
-	fprintf(out[index_ic],"%e\n",(psp_input->l[index_mode][psp_input->l_size[index_mode]-1]-1));
+	fprintf(out[index_ic],"%d\n",(int)(psp_input->l[index_mode][psp_input->l_size[index_mode]-1]-1));
         fprintf(out[index_ic],"#  l ");
 	for (index_ct=0; index_ct < psp_input->ct_size; index_ct++) {
 	  if ((ppt_input->has_cl_cmb_temperature == _TRUE_) &&
@@ -108,7 +108,7 @@ int output_init(
   }
 
   if (ppt_input->has_pk_matter == _TRUE_) {
-
+    
     if (pop->z_pk > ppr_input->z_max_pk) {
       sprintf(pop->error_message,"%s(L:%d) : P(k,z) computed up to z=%f but requested at z=%f. Must increase z_max_pk in precision file.",__func__,__LINE__,ppr_input->z_max_pk,pop->z_pk);
       return _FAILURE_;
@@ -118,7 +118,7 @@ int output_init(
 
     /* if z_pk = 0, no interpolation needed, 
        just let pk_output point to the right address */
-    if (pop->z_pk == 0.) {
+    if (pop->z_pk == 0) {
       pk_output=&(psp_input->pk[(psp_input->eta_size-1) * ppt_input->ic_size[index_mode] * psp_input->k_size]);
     }
